Modified test_forest to Pytest version
